Use Plek in the test_export_content test
As the Plek configuration isn't controlled by the test setup, this
test can break if the Plek configuration changes. So use Plek in the
test itself to make this work independently of the configuration.

This makes local development easier, as you can set the Plek
configuration, and not break the tests.