Rework the title and description tokenization
In the dataprep.py script.

Use the texts_to_matrix method, which removes the need to call the
texts_to_sequences and sequences_to_matrix methods in turn.

Also, set the num_words for the tokenizers prior to calling
texts_to_matrix, as this means that the size of the generated matrix
will be correct (and not have an excessive number of columns).