Fix tracing parent taxons and add taxon level into the output.
The previous 'explode' mechanism for tracing taxon parents relied
on the base_path for each taxon incuding all parent levels; this
is no longer the case - the taxons paths are now slugged. Instead,
the script now traces the parent hierarchy directly and defines
a 'topic taxon' as a 2-tuple of (level, base_path). Including the
level in the output will enable us to include it in the quality
measure.