Track contextual sidebar metrics
Track the amount of times each navigation type is displayed on the contextual
sidebar. Only one navigation type can appear and there is a priority in what
is displayed if the piece of content has more than one navigation type associated.

The list below is ordered in priority on which sidebar is displayed if
the piece of content has multiple navigation types;
1. Single step by step (step_by_step_nav)
2. Mainstream browse (related_navigation)
3. Curated related items (related_navigation)
4. Live taxon's (taxonomy_navigation)
5. Default(related_navigation)

Also track the total amount of content pages using each sidebar.

trello;
https://trello.com/c/fzL0r8Cs/158-ability-to-track-pages-using-the-new-taxonomy